extract  faq-content
