+++ app/common/help.tsx
@@ -1,0 +1,3 @@
+export default function HelpPage() {
+
+}
